requirements.txt
# Core dependencies
newsapi-python==0.2.7
requests>=2.32.3
python-dotenv==1.0.0
schedule==1.1.0
fastapi==0.104.1
uvicorn==0.24.0

# Development dependencies
pytest==7.4.0
